requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
boto3==1.28.72
botocore==1.31.72
bson==0.5.10
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.4
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
Naked==0.1.32
oauthlib==3.2.2
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
sendgrid==6.10.0
shellescape==3.8.1
six==1.16.0
sngrams==0.2.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tweet-parser==1.13.2
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
